requirements.txt
django==5.0.7
django-unicorn==0.60.0
psycopg==3.2.1
psycopg2_binary
haystack-ai==2.3.0
qdrant-haystack==4.1.2
ollama-haystack==0.0.7
pypdf==4.3.1
sentence-transformers>=2.3.0
transformers[torch,sentencepiece]
unstructured[pdf]